% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accmini}
\alias{accmini}
\title{A subset of the Adrenocortical Carcinoma (ACC) dataset}
\format{
A \code{FirehoseData} data object
}
\usage{
data("accmini", package = "RTCGAToolbox")
}
\description{
See the `acc_sample.R` script to see how the data was generated.
This dataset contains real data from the The Cancer Genome Atlas for
the pipeline run date and GISTIC analysis date of 2016-01-28.
}
\keyword{datasets}
