% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcRight.R
\name{bcRight}
\alias{bcRight}
\title{Extract TCGA barcode data}
\usage{
bcRight(barcodes)
}
\arguments{
\item{barcodes}{A character vector of TCGA barcodes}
}
\value{
A dataframe with sample type, sample code, vial, portion, analyte, plate, and center columns.
}
\description{
This function uses the TCGA barcode to return a data frame of the data to the right of the participant identifier.
}
\examples{

\dontrun{
barcodes <- colnames(getElement(a2, "RNASeqGene"))
right_bctable <- bcRight(barcodes)
}

}
\author{
Marcel Ramos
}

