% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocExtract.R
\name{biocExtract}
\alias{biocExtract}
\title{Extract data from \code{FirehoseData} object into \code{ExpressionSet} or
\code{GRangesList} object}
\usage{
biocExtract(object, type = c("Clinical", "RNASeqGene", "miRNASeqGene",
  "RNASeq2GeneNorm", "CNASNP", "CNVSNP", "CNASeq", "CNACGH", "Methylation",
  "Mutation", "mRNAArray", "miRNAArray", "RPPAArray", "GISTIC", "GISTICA",
  "GISTICT"), ...)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object,
see type section.}
}
\value{
Either an \linkS4class{SummarizedExperiment} object or a
\linkS4class{RaggedExperiment} object.
}
\description{
This function processes data from a \code{\link{FirehoseData}} object.
Raw data is converted to a conventional Bioconductor object. The function
returns either a \linkS4class{SummarizedExperiment} or a
\linkS4class{RaggedExperiment} class object. In cases where there are
multiple platforms in a data type, an attempt to consolidate datasets will
be made based on matching dimension names. For ranged data, this
functionality is provided with more control as part of the
\code{RaggedExperiment} features. See
\code{\link[RaggedExperiment]{RaggedExperiment}} for more details.
}
\section{type}{

Choices include: "RNAseqGene", "clinical", "miRNASeqGene",
"RNASeq2GeneNorm", "CNASNP", "CNVSNP", "CNASeq", "CNACGH", "Methylation",
"Mutation", "mRNAArray", "miRNAArray", "RPPAArray", "GISTIC", "GISTICA",
"GISTICT".
The "GISTICA" type of dataset represents GISTIC data by all
genes. "GISTICT" represents data thresholded by genes. To get both types
in a list, use "GISTIC".
}

\examples{
\dontrun{
library(RTCGAToolbox)
dataFolder <- normalizePath("~/Documents/data")
coadmut <- getFirehoseData("COAD", runDate = "20151101", Mutation = TRUE,
                         destdir = dataFolder)
cm <- biocExtract(coadmut, "Mutation")
}
}
\author{
Marcel Ramos \email{marcel.ramos@roswellpark.org}
}
