% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\name{Clinical}
\alias{Clinical}
\title{Get Clinical data from FirehoseData}
\usage{
Clinical(object)
}
\arguments{
\item{object}{}
}
\value{
Returns the Clinical data slot
}
\description{
Get Clinical data from FirehoseData
}
