% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{class}
\name{CorResult-class}
\alias{CorResult-class}
\title{An S4 class to store correlations between gene expression level and copy number data}
\description{
An S4 class to store correlations between gene expression level and copy number data
}
\section{Slots}{

\describe{
\item{\code{Dataset}}{A cohort name}

\item{\code{Correlations}}{Results data frame}
}}

