% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DGEResult-class}
\alias{DGEResult-class}
\title{An S4 class to store differential gene expression results}
\description{
An S4 class to store differential gene expression results
}
\section{Slots}{

\describe{
\item{\code{Dataset}}{Dataset name}

\item{\code{Toptable}}{Results data frame}
}}

