% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract function.}
\usage{
extract(object, type, phenoData = TRUE)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object.}

\item{phenoData}{Logical (default TRUE) includes clinic data if available.}
}
\value{
A \code{\link{ExpressionSet}} object for the selected data type.
}
\description{
\code{extract} returns an S4 ExpressionSet object from "FirehoseData" object.
}
\details{
This function serves to extract and reorganize data into a structured S4
object for genomic analysis. An option is available to retreive additional
clinical data and include it as featureData. See \code{\link{ExpressionSet}}
}
\examples{
\dontrun{
b2 <- extract(a2, "Methylation", phenoData=TRUE)
}
}

