% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{class}
\name{FirehoseCGHArray-class}
\alias{FirehoseCGHArray-class}
\title{An S4 class to store data from CGA platforms}
\description{
An S4 class to store data from CGA platforms
}
\section{Slots}{

\describe{
\item{\code{Filename}}{Platform name}

\item{\code{DataMatrix}}{A data frame that stores the CGH data.}
}}

