% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firehose-extractors.R
\name{FirehoseData-extractors}
\alias{FirehoseData-extractors}
\alias{clinical}
\alias{RNASeqGene}
\alias{RNASeq2GeneNorm}
\alias{miRNASeqGene}
\alias{CNASNP}
\alias{CNVSNP}
\alias{CNASeq}
\alias{CNACGH}
\alias{Methylation}
\alias{mRNAArray}
\alias{miRNAArray}
\alias{RPPAArray}
\alias{Mutation}
\alias{GISTIC}
\title{Accessor methods for the FirehoseData object}
\usage{
clinical(object)

RNASeqGene(object)

RNASeq2GeneNorm(object)

miRNASeqGene(object)

CNASNP(object)

CNVSNP(object)

CNASeq(object)

CNACGH(object)

Methylation(object)

mRNAArray(object)

miRNAArray(object)

RPPAArray(object)

Mutation(object)

GISTIC(object)
}
\arguments{
\item{object}{A \code{FirehoseData} class object}
}
\description{
A list of accessor functions for the \linkS4class{FirehoseData}
class. Each exposing a unique element of the \linkS4class{FirehoseData}
object. See \link{FirehoseData-class} for more details.

\itemize{
    \item{clinical} - Get the clinical data slot
    \item{RNASeqGene} - RNASeqGene
    \item{RNASeq2GeneNorm} - Normalized
    \item{miRNASeqGene} - micro RNA SeqGene
    \item{CNASNP} - Copy Number Alteration
    \item{CNVSNP} - Copy Number Variation
    \item{CNASeq} - Copy Number Alteration
    \item{CNACGH} - Copy Number Alteration
    \item{Methylation} - Methylation
    \item{mRNAArray} - Messenger RNA
    \item{miRNAArray} - micro RNA
    \item{RPPAArray} - Reverse Phase Protein Array
    \item{Mutation} - Mutations
    \item{GISTIC} - GISTIC v2 scores and probabilities
}
}
