% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FirehoseMethylationArray-class}
\alias{FirehoseMethylationArray-class}
\title{An S4 class to store data from methylation platforms}
\description{
An S4 class to store data from methylation platforms
}
\section{Slots}{

\describe{
\item{\code{Filename}}{Platform name}

\item{\code{DataMatrix}}{A data frame that stores the methylation data.}
}}

