% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCNGECorrelation.R
\name{getCNGECorrelation}
\alias{getCNGECorrelation}
\title{Perform correlation analysis betwwen gene expression and copy number data}
\usage{
getCNGECorrelation(dataObject, adj.method = "BH", adj.pval = 0.05,
  raw.pval = 0.05)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{adj.method}{Raw p value adjustment methods (Default "BH")}

\item{adj.pval}{Adjusted p value cut off for results table (Default 0.05)}

\item{raw.pval}{raw p value cut off for results table (Default 0.05)}
}
\value{
Returns a list that stores results for each dataset
}
\description{
\code{getCNGECorrelation} returns a list that stores the results correlation between gene expression and copy number data.
}
\examples{
data(RTCGASample)
corRes = getCNGECorrelation(RTCGASample)
corRes
showResults(corRes[[1]])
\dontrun{
}
}

