% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{methods}
\name{getData,FirehoseData-method}
\alias{getData,FirehoseData-method}
\alias{getData,FirehoseData,FirehoseData-method}
\title{Export data from FirehoseData object}
\usage{
\S4method{getData}{FirehoseData}(object, type = "", platform = NULL,
  CN = "All")
}
\arguments{
\item{object}{A \code{\linkS4class{FirehoseData}} object}

\item{type}{A data type to be exported (Data types can be seen by typing show(objectname))}

\item{platform}{A list id for data types that may come from multiple platform (such as mRNAArray)}

\item{CN}{A copy number data type (Default: 'All') (Possible values 'All' or 'Thresholded')}
}
\value{
Returns matrix or data frame depends on data type
}
\description{
Export data from FirehoseData object
}
\examples{
data(RTCGASample)
sampleClinical = getData(RTCGASample,"clinical")
sampleClinical = getData(RTCGASample,"RNASeqGene")
}
