% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{methods}
\name{getData}
\alias{getData}
\title{Export data from FirehoseData object}
\usage{
getData(object, type, platform)

\S4method{getData}{FirehoseData}(object, type = "clinical", platform)
}
\arguments{
\item{object}{A \code{\linkS4class{FirehoseData}} object}

\item{type}{A data type to be exported (default "clinical")}

\item{platform}{An index for data types that may come from multiple
platforms (such as mRNAArray), for GISTIC data, one of the options:
'AllByGene' or 'ThresholdedByGene'}
}
\value{
Returns matrix or data.frame depending on data type
}
\details{
Available datatypes for a particular object can be seen by
entering the object name in the console ('show' method)
}
\section{Methods (by class)}{
\itemize{
\item \code{FirehoseData}: Get a matrix or data.frame from \code{FirehoseData}
}}

\examples{
data(RTCGASample)
getData(RTCGASample)
getData(RTCGASample, "RNASeqGene")

}
