% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirehoseAnalyzeDates.R
\name{getFirehoseAnalyzeDates}
\alias{getFirehoseAnalyzeDates}
\title{Get data analyze dates.}
\usage{
getFirehoseAnalyzeDates(last = NULL)
}
\arguments{
\item{last}{To list last n dates. (Default NULL)}
}
\value{
A character vector for dates.
}
\description{
\code{getFirehoseAnalyzeDates} returns the character vector for analyze release dates.
}
\examples{
getFirehoseAnalyzeDates()
getFirehoseAnalyzeDates(last=2)
}

