% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFirehoseDatasets}
\alias{getFirehoseDatasets}
\title{Get list of TCGA cohorts.}
\usage{
getFirehoseDatasets()
}
\value{
A character string
}
\description{
\code{getFirehoseDatasets} returns a character array for cohorts.
}
\examples{
getFirehoseDatasets()
}

