% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFirehoseRunningDates.R
\name{getFirehoseRunningDates}
\alias{getFirehoseRunningDates}
\title{Get standard data running dates.}
\usage{
getFirehoseRunningDates(last = NULL)
}
\arguments{
\item{last}{To list last n dates. (Default NULL)}
}
\value{
A character vector for dates.
}
\description{
\code{getFirehoseRunningDates} returns the character vector for standard data release dates.
}
\examples{
getFirehoseRunningDates()
getFirehoseRunningDates(last=2)
}

