% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGISTICPeaks.R
\name{getGISTICPeaks}
\alias{getGISTICPeaks}
\title{Download peak-level data from the Firehose pipeline}
\usage{
getGISTICPeaks(dataset, peak = c("wide", "narrow", "full"), rm.chrX = TRUE,
  gistic2Date = "20160128")
}
\arguments{
\item{dataset}{A TCGA cancer code}

\item{peak}{The peak type, select from "wide", "narrow", "full".}

\item{rm.chrX}{(logical default TRUE) Whether to remove observations in the
X chromosome}

\item{gistic2Date}{(character default "20160128") Data of the analysis
pipeline run}
}
\value{
A \code{data.frame} of peak values
}
\description{
Access copy number GISTIC2 level 4 data through
\code{gdac.broadinstitute.org}
}
\examples{

co <- getGISTICPeaks("COAD", "wide")
class(co)
head(co)[1:6]

}
\author{
Ludwig Geistlinger
}
