% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{DEFUNCT: Draws a circle plot into working directory}
\usage{
getReport(dataObject, DGEResult1 = NULL, DGEResult2 = NULL, geneLocations)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{DGEResult1}{Differential gene expression results object (Optional)}

\item{DGEResult2}{Differential gene expression results object (Optional)}

\item{geneLocations}{Gene coordinates.}
}
\value{
Draws a circle plot
}
\description{
\code{getReport} draws a circle plot into your working directory to show log
fold changes for differentially expressed genes, copy number alterations and
mutations.
}
\examples{
data(accmini)
}
