% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hg19.ucsc.gene.locations}
\alias{hg19.ucsc.gene.locations}
\title{Gene coordinates for circle plot.}
\format{A data frame with 28454 rows and 5 variables}
\description{
A dataset containing the gene coordinates
The variables are as follows:
}
\details{
\itemize{
  \item GeneSymbol. Gene symbols
  \item Chromosome. Chromosome name
  \item Strand. Gene strand on chromosome
  \item Start. Gene location on chromosome
  \item End. Gene location on chromosome
}
}

