% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RTCGASample}
\alias{RTCGASample}
\title{A sample data object for sample codes.}
\format{A \code{FirehoseData} data object}
\description{
A \code{FirehoseData} object for running sample codes.
The variables are as follows:
}
\details{
\itemize{
  \item a2. A sample data object
}
}

