% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-defunct.R
\name{RTCGAToolbox-defunct}
\alias{RTCGAToolbox-defunct}
\alias{getCNGECorrelation}
\alias{getDiffExpressedGenes}
\alias{getSurvival}
\title{Functions that are defunct in the package.}
\usage{
getCNGECorrelation(
  dataObject,
  adj.method = "BH",
  adj.pval = 0.05,
  raw.pval = 0.05
)

getDiffExpressedGenes(
  dataObject,
  DrawPlots = TRUE,
  adj.method = "BH",
  adj.pval = 0.05,
  raw.pval = 0.05,
  logFC = 2,
  hmTopUpN = 100,
  hmTopDownN = 100,
  meanFilter = 10
)

getSurvival(dataObject, numberofGroups = 2, geneSymbols, sampleTimeCensor)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{adj.method}{Raw p value adjustment methods (Default "BH")}

\item{adj.pval}{Adjusted p value cut off for results table (Default 0.05)}

\item{raw.pval}{raw p value cut off for results table (Default 0.05)}

\item{DrawPlots}{A logical parameter to draw heatmaps and volcano plots.}

\item{logFC}{log fold change cut off for results table (Default 2)}

\item{hmTopUpN}{Max number of up regulated genes in heatmap (Default 100)}

\item{hmTopDownN}{Max number of down regulated genes in heatmap (Default 100)}

\item{meanFilter}{Mean read counts for each gene to filter not expressed
genes (Default 10)}

\item{numberofGroups}{Can be set as 2 or 3. (Default 2) Order and divide samples into n groups by using gene expression data.}

\item{geneSymbols}{Gene symbol that is going to be tested}

\item{sampleTimeCensor}{a data frame that stores clinical data. First column should store sample IDs, second column should have time and third column should have event information. For more information please see vignette.}
}
\value{
Returns a list that stores results for each dataset

getDiffExpressedGenes: Returns a list that stores results for each
  dataset

getSurvival: Draws a KM plot
}
\description{
\code{getCNGECorrelation}: Perform correlation analysis between
  gene expression and copy number data. \code{getDiffExpressedGenes}: Perform
  differential gene expression analysis for mRNA expression data.
  \code{getSurvival}: Perform survival analysis based on gene expression
  data.
}
\details{
\code{getCNGECorrelation} returns a list that stores the results correlation
between gene expression and copy number data.
\code{getDiffExpressedGenes} returns a list that stores the results for each
dataset.
\code{getSurvival} draws a KM plot and show survival analysis results between
groups that are defined by gene expression data
}
