\name{RTCGAToolbox-package}
\alias{RTCGAToolbox-package}
\alias{RTCGAToolbox}
\docType{package}
\title{
RTCGAToolbox: A data client and basic analysis toolbox for Firehose TCGA data.
}
\description{
Managing data from large scale projects such as The Cancer Genome Atlas (TCGA) for further analysis is an important and time consuming step for research projects. Several efforts, such as Firehose project, make TCGA pre-processed data publicly available via web services and data portals but it requires managing, downloading and preparing the data for following steps. We developed an open source and extensible R based data client for Firehose pre-processed data and demonstrated its use with sample case studies.
}
\details{
\tabular{ll}{
Package: \tab RTCGAToolbox\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-19\cr
License: \tab GPL (>=2) \cr
Depends: \tab methods,XML,limma,survival,RCircos\cr
}

}
\author{
Mehmet Kemal Samur

Maintainer: Mehmet Kemal Samur <samur@jimmy.harvard.edu>

}
\keyword{ package }
