% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firehose-extractors.R
\name{selectType}
\alias{selectType}
\title{Accessor function for the FirehoseData object}
\usage{
selectType(object, dataType)
}
\arguments{
\item{object}{A \code{FirehoseData} class object}

\item{dataType}{A data type, see details.}
}
\value{
The data type element of the \code{FirehoseData} object
}
\description{
An accessor function for the \linkS4class{FirehoseData}
class. An argument will specify the data type to return
See \link{FirehoseData-class} for more details.
}
\details{
\itemize{
    \item{clinical} - Get the clinical data slot
    \item{RNASeqGene} - RNASeqGene
    \item{RNASeq2GeneNorm} - Normalized
    \item{miRNASeqGene} - micro RNA SeqGene
    \item{CNASNP} - Copy Number Alteration
    \item{CNVSNP} - Copy Number Variation
    \item{CNASeq} - Copy Number Alteration
    \item{CNACGH} - Copy Number Alteration
    \item{Methylation} - Methylation
    \item{mRNAArray} - Messenger RNA
    \item{miRNAArray} - micro RNA
    \item{RPPAArray} - Reverse Phase Protein Array
    \item{Mutation} - Mutations
    \item{GISTIC} - GISTIC v2 scores and probabilities
}
}
