% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\name{showResults,CorResult-method}
\alias{showResults,CorResult-method}
\alias{showResults,CorResult,CorResult-method}
\title{Export toptable or correlation data frame}
\usage{
\S4method{showResults}{CorResult}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{DGEResult}} or \code{\linkS4class{CorResult}} object}
}
\value{
Returns correlation results data frame
}
\description{
Export toptable or correlation data frame
}
\examples{
data(accmini)
}
