% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\name{showResults,DGEResult-method}
\alias{showResults,DGEResult-method}
\alias{showResults,DGEResult,DGEResult-method}
\title{Export toptable or correlation data frame}
\usage{
\S4method{showResults}{DGEResult}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{DGEResult}} or \code{\linkS4class{CorResult}} object}
}
\value{
Returns toptable for DGE results
}
\description{
Export toptable or correlation data frame
}
\examples{
data(RTCGASample)
dgeRes = getDiffExpressedGenes(RTCGASample)
dgeRes
showResults(dgeRes[[1]])
}
