% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\name{showResults}
\alias{showResults}
\title{Export toptable or correlation data frame}
\usage{
showResults(object)
}
\arguments{
\item{object}{A \code{\linkS4class{DGEResult}} or \code{\linkS4class{CorResult}} object}
}
\value{
Returns toptable or correlation data frame
}
\description{
Export toptable or correlation data frame
}
\examples{
data(accmini)
}
