% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocExtract.R
\name{biocExtract}
\alias{biocExtract}
\title{Extract and convert data from a \code{FirehoseData} object to a
\code{Bioconductor} object}
\usage{
biocExtract(
  object,
  type = c("clinical", "RNASeqGene", "miRNASeqGene", "RNASeq2GeneNorm", "CNASNP",
    "CNVSNP", "CNASeq", "CNACGH", "Methylation", "Mutation", "mRNAArray", "miRNAArray",
    "RPPAArray", "GISTIC", "GISTICA", "GISTICT", "GISTICP")
)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object,
see type section.}
}
\value{
Either an \linkS4class{SummarizedExperiment} object or a
\linkS4class{RaggedExperiment} object.
}
\description{
This function processes data from a \code{\linkS4class{FirehoseData}} object.
Raw data is converted to a conventional Bioconductor object. The function
returns either a \linkS4class{SummarizedExperiment} or a
\linkS4class{RaggedExperiment} class object. In cases where there are
multiple platforms in a data type, an attempt to consolidate datasets will
be made based on matching dimension names. For ranged data, this
functionality is provided with more control as part of the
\code{RaggedExperiment} features. See
\code{\link[RaggedExperiment]{RaggedExperiment}} for more details.
}
\section{type}{

Choices include:
\itemize{
    \item{clinical} - Get the clinical data slot
    \item{RNASeqGene} - RNASeqGene
    \item{RNASeq2GeneNorm} - Normalized
    \item{miRNASeqGene} - micro RNA SeqGene
    \item{CNASNP} - Copy Number Alteration
    \item{CNVSNP} - Copy Number Variation
    \item{CNASeq} - Copy Number Alteration
    \item{CNACGH} - Copy Number Alteration
    \item{Methylation} - Methylation
    \item{mRNAArray} - Messenger RNA
    \item{miRNAArray} - micro RNA
    \item{RPPAArray} - Reverse Phase Protein Array
    \item{Mutation} - Mutations
    \item{GISTICA} - GISTIC v2 ('AllByGene' only)
    \item{GISTICT} - GISTIC v2 ('ThresholdedByGene' only)
    \item{GISTICP} - GISTIC v2 ('Peaks' only)
    \item{GISTIC} - GISTIC v2 scores, probabilities, and peaks
}
}

\examples{
\dontrun{
    coadmut <- getFirehoseData("COAD", Mutation = TRUE)
    biocExtract(coadmut, "Mutation")
}
}
\author{
Marcel Ramos \email{marcel.ramos@roswellpark.org}
}
