% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract data from \code{FirehoseData} object into \code{ExpressionSet} or
\code{GRangesList} object}
\usage{
extract(object, type = NULL)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object.}
}
\description{
This extracts data from a \code{FirehoseData} object and converts it to a
structured S4 object for analysis. The function returns an ExpressionSet or
GRangesList #' class object (see \code{\link{ExpressionSet}} and
\code{\link{GRangesList}}).
}
\examples{

\dontrun{
b2 <- extract(a2, "Methylation", clinical=TRUE)
}

}
\author{
Marcel Ramos \email{mramos09@gmail.com}
}

