% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract data from \code{FirehoseData} object into S4 \code{ExpressionSet}}
\usage{
extract(object, type, phenoData = TRUE)
}
\arguments{
\item{object}{A \code{FirehoseData} object from which to extract data.}

\item{type}{The type of data to extract from the "FirehoseData" object.}

\item{phenoData}{Logical (default TRUE) includes additional clinic data, if available.}
}
\value{
An \code{\link{ExpressionSet}} object for the selected data type.
Choices include: "RNAseq_Gene", "Clinic", "miRNASeq_Gene", "RNAseq2_Gene_Norm", "CNA_SNP", "CNV_SNP", "CNA_Seq", "CNA_CGH", "Methylation", "Mutation", "mRNA_Array", "miRNA_Array", "RPPA", "GISTIC_A", "GISTIC_T".
The "GISTIC_A" type of dataset represents GISTIC data by all genes. "GISTIC_T"" represents data thresholded by genes.
}
\description{
This function serves to extract and reorganize data into a structured S4
object for genomic analysis. An option is available to retreive additional
clinical data and include it as featureData. See \code{\link{ExpressionSet}}
}
\examples{
\dontrun{
b2 <- extract(a2, "Methylation", phenoData=TRUE)
}
}

