% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R, R/Firehose-extractors.R
\docType{class}
\name{FirehoseData-class}
\alias{FirehoseData-class}
\alias{show,FirehoseData-method}
\alias{updateObject,FirehoseData-method}
\title{An S4 class to store main data object from clinent function.}
\usage{
\S4method{show}{FirehoseData}(object)

\S4method{updateObject}{FirehoseData}(object, ..., verbose = FALSE)

\S4method{selectType}{FirehoseData}(object, dataType)
}
\arguments{
\item{object}{A FirehoseData object}

\item{...}{additional arguments for updateObject}

\item{verbose}{logical (default FALSE) whether to print extra messages}
}
\description{
An S4 class to store main data object from clinent function.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: show method

\item \code{updateObject}: Update an old RTCGAToolbox FirehoseData object to
the most recent API

\item \code{selectType}: Extract data type
}}

\section{Slots}{

\describe{
\item{\code{Dataset}}{A cohort name}

\item{\code{runDate}}{Standard data run date from \code{\link{getFirehoseRunningDates}}}

\item{\code{gistic2Date}}{Analyze running date from \code{\link{getFirehoseAnalyzeDates}}}

\item{\code{clinical}}{clinical data frame}

\item{\code{RNASeqGene}}{Gene level expression data matrix from RNAseq}

\item{\code{RNASeq2GeneNorm}}{Gene level expression data matrix from RNAseq (RSEM)}

\item{\code{miRNASeqGene}}{miRNA expression data from matrix smallRNAseq}

\item{\code{CNASNP}}{A data frame to store somatic copy number alterations from SNP array platform}

\item{\code{CNVSNP}}{A data frame to store germline copy number variants from SNP array platform}

\item{\code{CNASeq}}{A data frame to store somatic copy number alterations from sequencing platform}

\item{\code{CNACGH}}{A list that stores \code{FirehoseCGHArray} object for somatic copy number alterations from CGH platform}

\item{\code{Methylation}}{A list that stores \code{FirehoseMethylationArray} object for methylation data}

\item{\code{mRNAArray}}{A list that stores \code{FirehosemRNAArray} object for gene expression data from microarray}

\item{\code{miRNAArray}}{A list that stores \code{FirehosemRNAArray} object for miRNA expression data from microarray}

\item{\code{RPPAArray}}{A list that stores \code{FirehosemRNAArray} object for RPPA data}

\item{\code{Mutation}}{A data frame for mutation infromation from sequencing data}

\item{\code{GISTIC}}{A \code{FirehoseGISTIC} object to store processed copy number data}

\item{\code{BarcodeUUID}}{A data frame that stores the Barcodes, UUIDs and Short sample identifiers}
}}

