% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{class}
\name{FirehoseGISTIC-class}
\alias{FirehoseGISTIC-class}
\title{An S4 class to store processed copy number data. (Data processed by using GISTIC2 algorithm)}
\description{
An S4 class to store processed copy number data. (Data processed by using GISTIC2 algorithm)
}
\section{Slots}{

\describe{
\item{\code{Dataset}}{Cohort name}

\item{\code{AllByGene}}{A data frame that stores continuous copy number}

\item{\code{ThresholdedByGene}}{A data frame for discrete copy number data}
}}

