% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{class}
\name{FirehoseMethylationArray-class}
\alias{FirehoseMethylationArray-class}
\title{An S4 class to store data from methylation platforms}
\description{
An S4 class to store data from methylation platforms
}
\section{Slots}{

\describe{
\item{\code{Filename}}{Platform name}

\item{\code{DataMatrix}}{A data frame that stores the methylation data.}
}}

