% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RTCGAToolbox-Class.R
\docType{class}
\name{FirehosemRNAArray-class}
\alias{FirehosemRNAArray-class}
\title{An S4 class to store data from array (mRNA, miRNA etc.) platforms}
\description{
An S4 class to store data from array (mRNA, miRNA etc.) platforms
}
\section{Slots}{

\describe{
\item{\code{Filename}}{Platform name}

\item{\code{DataMatrix}}{A data matrix that stores the expression data.}
}}

