% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBroadSubtypes.R
\name{getBroadSubtypes}
\alias{getBroadSubtypes}
\title{Download Broad Institute Subtype Clustering Results}
\usage{
getBroadSubtypes(dataset, clust.alg = c("CNMF", "ConsensusPlus"))
}
\arguments{
\item{dataset}{A TCGA cancer code}

\item{clust.alg}{The selected cluster algorithm, either "CNMF" or
"ConsensuPlus"}
}
\value{
A \code{data.frame} of cluster and silhouette values
}
\description{
Obtain the subtype clustering results from the Broad for
a specific cancer code (see \link{getFirehoseDatasets}).
}
\examples{
co <- getBroadSubtypes("COAD", "CNMF")
head(co)

}
\author{
Ludwig Geistlinger
}
