% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDiffExpressedGenes}
\alias{getDiffExpressedGenes}
\title{Perform differential gene expression analysis for mRNA expression data.}
\usage{
getDiffExpressedGenes(dataObject, DrawPlots = TRUE, adj.method = "BH",
  adj.pval = 0.05, raw.pval = 0.05, logFC = 2, hmTopUpN = 100,
  hmTopDownN = 100, meanFilter = 10)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{DrawPlots}{A logical parameter to draw heatmaps and volcano plots.}

\item{adj.method}{Raw p value adjustment methods (Default "BH")}

\item{adj.pval}{Adjusted p value cut off for results table (Default 0.05)}

\item{raw.pval}{raw p value cut off for results table (Default 0.05)}

\item{logFC}{log fold change cut off for results table (Default 2)}

\item{hmTopUpN}{Max number of up regulated genes in heatmap (Default 100)}

\item{hmTopDownN}{Max number of down regulated genes in heatmap (Default 100)}

\item{meanFilter}{Mean read counts for each gene to filter not expressed genes (Default 10)}
}
\value{
Returns a list that stores results for each dataset
}
\description{
\code{getDiffExpressedGenes} returns a list that stores the results for each dataset.
}
\examples{
\dontrun{
data(RTCGASample)
dgegenes = getDiffExpressedGenes(a2)
}
}

