% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFirehoseData}
\alias{getFirehoseData}
\title{Get data from Firehose portal.}
\usage{
getFirehoseData(dataset, runDate = NULL, gistic2_Date = NULL,
  RNAseq_Gene = FALSE, Clinic = TRUE, miRNASeq_Gene = FALSE,
  RNAseq2_Gene_Norm = FALSE, CNA_SNP = FALSE, CNV_SNP = FALSE,
  CNA_Seq = FALSE, CNA_CGH = FALSE, Methylation = FALSE,
  Mutation = FALSE, mRNA_Array = FALSE, miRNA_Array = FALSE,
  RPPA = FALSE, RNAseqNorm = "raw_counts",
  RNAseq2Norm = "normalized_count", todir = NULL, forceDownload = FALSE,
  fileSizeLimit = 500, getUUIDs = FALSE)
}
\arguments{
\item{dataset}{A cohort name. All dataset names can be accessible via \code{\link{getFirehoseDatasets}}}

\item{runDate}{Standard data run dates. Date list can be accessible via \code{\link{getFirehoseRunningDates}}}

\item{gistic2_Date}{Analyze running dates for GISTIC processed copy number data. Date list can be accessible via \code{\link{getFirehoseAnalyzeDates}}}

\item{RNAseq_Gene}{Logical (default FALSE) parameter for RNAseq data.}

\item{Clinic}{Logical (default TRUE) parameter for clinical data.}

\item{miRNASeq_Gene}{Logical (default FALSE) parameter for smallRNAseq data.}

\item{RNAseq2_Gene_Norm}{Logical (default FALSE) parameter for RNAseq v2 (RSEM processed) data.}

\item{CNA_SNP}{Logical (default FALSE) parameter for somatic copy number alterations data from SNP array.}

\item{CNV_SNP}{Logical (default FALSE) parameter for germline copy number variants data from SNP array.}

\item{CNA_Seq}{Logical (default FALSE) parameter for somatic copy number alterations data from sequencing.}

\item{CNA_CGH}{Logical (default FALSE) parameter for somatic copy number alterations data from CGH.}

\item{Methylation}{Logical (default FALSE) parameter for methylation data.}

\item{Mutation}{Logical (default FALSE) parameter for mutation data from sequencing.}

\item{mRNA_Array}{Logical (default FALSE) parameter for mRNA expression data from microarray.}

\item{miRNA_Array}{Logical (default FALSE) parameter for miRNA expression data from microarray.}

\item{RPPA}{Logical (default FALSE) parameter for RPPA data}

\item{RNAseqNorm}{RNAseq data normalization method. (Default raw_counts)}

\item{RNAseq2Norm}{RNAseq v2 data normalization method. (Default normalized_count)}

\item{todir}{The directory to download files to and from which to read files. It will be created if necessary. Defaults to current working directory.}

\item{forceDownload}{A logic (Default FALSE) key to force download files when run. By default, local files in the specified directory will be used if available.}

\item{fileSizeLimit}{Files that are larger than set value (megabyte) won't be downloaded (Default: 500)}

\item{getUUIDs}{Logical key to get UUIDs from barcode (Default: FALSE)}
}
\value{
A \code{FirehoseData} data object that stores data for selected data types.
}
\description{
\code{getFirehoseData} returns \code{FirehoseData} object that stores TCGA data.
}
\details{
This is a main client function to download data from Firehose TCGA portal.
}
\examples{
\dontrun{
a2 <- getFirehoseData(dataset="BRCA",
runDate="20140416",gistic2_Date="20140115",
RNAseq_Gene=TRUE,Clinic=TRUE,mRNA_Array=TRUE,Mutation=TRUE)
}
}

