% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirehoseDatasets.R
\name{getFirehoseDatasets}
\alias{getFirehoseDatasets}
\title{Get a list of TCGA disease cohorts}
\usage{
getFirehoseDatasets()
}
\value{
A character string
}
\description{
\code{getFirehoseDatasets} returns a character vector of TCGA disease codes.
A reference table can be seen at \url{https://gdac.broadinstitute.org/}.
}
\examples{

getFirehoseDatasets()

}
\seealso{
\url{https://gdac.broadinstitute.org/}
}
