% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFirehoseDatasets.R
\name{getFirehoseDatasets}
\alias{getFirehoseDatasets}
\title{Get list of TCGA cohorts.}
\usage{
getFirehoseDatasets()
}
\value{
A character string
}
\description{
\code{getFirehoseDatasets} returns a character array for cohorts.
}
\examples{
getFirehoseDatasets()
}

