% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutationRate.R
\name{getMutationRate}
\alias{getMutationRate}
\title{Make a table for mutation rate of each gene in the cohort}
\usage{
getMutationRate(dataObject)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}
}
\value{
Returns a data table
}
\description{
Make a table for mutation rate of each gene in the cohort
}
\examples{
data(RTCGASample)
mutRate <- getMutationRate(dataObject=RTCGASample)
mutRate <- mutRate[order(mutRate[,2],decreasing = TRUE),]
head(mutRate)
\dontrun{
}
}
