% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{Draws a circle plot into working directory}
\usage{
getReport(dataObject, DGEResult1 = NULL, DGEResult2 = NULL, geneLocations)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{DGEResult1}{Differential gene expression results object (Optional)}

\item{DGEResult2}{Differential gene expression results object (Optional)}

\item{geneLocations}{Gene coordinates.}
}
\value{
Draws a circle plot
}
\description{
\code{getReport} draws a circle plot into your workin director to show log fold changes for differentially expressed genes, copy number alterations and mutations.
}
\examples{
data(RTCGASample)
data(hg19.ucsc.gene.locations)
t1=getDiffExpressedGenes(a2)
\dontrun{
getReport(dataObject=a2,DGEResult1=t1[[1]],geneLocations=hg19.ucsc.gene.locations)
}
}

