% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{Draws a circle plot into working directory}
\usage{
getReport(dataObject, DGEResult1 = NULL, DGEResult2 = NULL, geneLocations)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{DGEResult1}{Differential gene expression results object (Optional)}

\item{DGEResult2}{Differential gene expression results object (Optional)}

\item{geneLocations}{Gene coordinates.}
}
\value{
Draws a circle plot
}
\description{
\code{getReport} draws a circle plot into your workin director to show log fold changes for differentially expressed genes, copy number alterations and mutations.
}
\examples{
data(accmini)
require("Homo.sapiens")
locations <- genes(Homo.sapiens,columns="SYMBOL")
locations <- as.data.frame(locations)
locations <- locations[,c(6,1,5,2:3)]
locations <- locations[!is.na(locations[,1]),]
locations <- locations[!duplicated(locations[,1]),]
rownames(locations) <- locations[,1]
t1 <- getDiffExpressedGenes(accmini)
\dontrun{
getReport(dataObject=accmini,DGEResult1=t1[[1]],geneLocations=locations)
}
}
