% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSurvival.R
\name{getSurvival}
\alias{getSurvival}
\title{Perform survival analysis based on gene expression data}
\usage{
getSurvival(dataObject, numberofGroups = 2, geneSymbols, sampleTimeCensor)
}
\arguments{
\item{dataObject}{This must be \code{FirehoseData} object.}

\item{numberofGroups}{Can be set as 2 or 3. (Default 2) Order and divide samples into n groups by using gene expression data.}

\item{geneSymbols}{Gene symbol that is going to be tested}

\item{sampleTimeCensor}{a data frame that stores clinical data. First column should store sample IDs, second column should have time and third column should have event information. For more information please see vignette.}
}
\value{
Draws a KM plot
}
\description{
\code{getSurvival} draws a KM plot and show survival analysis results between groups that are defined by gene expression data
}
\examples{
\dontrun{
clinicData = brcaData@Clinical
clinicData = clinicData[,3:5]
clinicData[is.na(clinicData[,3]),3] = clinicData[is.na(clinicData[,3]),2]
survData <- data.frame(Samples=rownames(clinicData),Time=as.numeric(clinicData[,3]),
Censor=as.numeric(clinicData[,1]))
getSurvival(dataObject=brcaData,geneSymbols=c("PIK3CA"),sampleTimeCensor=survData)
}
}

