% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGAToolbox.R, R/getFirehoseDatasets.R
\docType{data}
\name{RTCGASample}
\alias{RTCGASample}
\title{A sample data object for sample codes.}
\format{A \code{FirehoseData} data object}
\usage{
RTCGASample
}
\description{
A \code{FirehoseData} object for running sample codes.
The variables are as follows:

Example dataset not biologically meaningful
}
\details{
\itemize{
  \item a2. A sample data object
}
}
\keyword{data}
